requirements.txt
channels==4.0.0
daphne==4.0.0
Django==4.1.7
requests==2.28.2
python-decouple==3.6
sentry-sdk==1.17.0